requirements.txt
django>=1.8,<1.9
django-crispy-forms
django-model-utils
psycopg2==2.5
